+++ tests/consoleTests.js
@@ -9,13 +9,10 @@ exports.consoleLog = function(test) {
         lastWrite = string;
     };
     log.instance = new log('debug');
-    log.instance.getDateString = function() {
-        return 'date';
-    };
     log.setLevel('debug');
     log.debug('test');
     //no \n is printed when using console.log
-    test.equal(lastWrite, '~ [date] DEBUG -- test');
+    test.equal(lastWrite, '~ DEBUG -- test');
     console.log = oldLog;
     delete global.LLOG_SKIP_USING_PROCESS;
     test.done();